requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.11.0
google-generativeai==0.3.0
